github-trend/python/requirements.txt
beautifulsoup4
requests
GitPython